[UPDATE]: Add Org-mode and friends to workflow!
This commit fixes several issues with the previous config. Firstly,
emojis package has been added to ensure no characters show up as
"tofus" any longer! Secondly, ivy, swiper, ivy-rich, and counsel
configurations have been reworked to ensure the loading is done
properly and not lead to errors any longer. Third, as the heading
suggests, we are now using org-mode finally! Literate programming here
I come! Besides that are a few additions to shortcuts added and as
some minor fixes in the config for go. We have added a dashboard as
well as a unobtrusive way to view pairs of parenthesis that are out of
the page!